requirements.txt
asgiref==3.7.2
dj-database-url==2.2.0
Django==5.0.3
django-filter==24.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
